getsubopt(), dd, misc
fixed VBE and VGA potentially causing a massive OOB write
fixed malloc not returning ENOMEM
fixed time syscall returning ENOSYS whilst returning time
fixed tty handling causing \r to do \n
fixed stderr being line buffered
fixed signal dispatching nuking old syscall return values
fixed the symbol table and the string table being marked as free

retyped rand() to be posix compliant
included a couple of headers required to be in stdlib.h

+++ libc/src/stdlib.c
@@ -11,7 +11,7 @@
 
 static uint32_t ___internal_rand_state = 1;
 
-uint32_t rand() {
+int rand() {
     if (__builtin_expect(___internal_rand_state == 0, 0))
         ___internal_rand_state = 0xDEADBEEF;
 
@@ -21,7 +21,7 @@ uint32_t rand() {
     ___internal_rand_state ^= ___internal_rand_state >> 17;
     ___internal_rand_state ^= ___internal_rand_state << 5;
 
-    return ___internal_rand_state % RAND_MAX;
+    return (int)(___internal_rand_state % RAND_MAX);
 }
 
 void srand(uint32_t seed) {___internal_rand_state = seed;}
